refactor: Transition mounting logic to behavior-based composition
